prime now supports ES5 property descriptors.
the mutator method has been removed in favor of the define method. the define method always takes a descriptor as the second argument, and works like Object.defineProperty, except on the prototype of the prime.
In environments without defineProperty support, each property is simply being set to the prototype.
prime.define is the generic define function, which uses ES5's Object.defineProperty where available, otherwise simply sets the descriptor's value to the object.
If an implemented property isnt an own property of the passed object (implementing a prototype that inherits stuff for instance) then prime will create another descriptor with default values.