updated ghost & shell for the new prime define.
ghost is largely simplified. registering a ghost now involves mostly automatic type detection.
ghost types are now saved to ghost.
ghost does not override the base define/mutator any longer, since a ghost is made of a snapshot of the current prototype chain of the base object. rather, when implementing in a specific ghost, the base object is updated as well.